requirements.txt
google-generativeai
yfinance
pandas
requests